+++ packages/ui/certd-client/src/locales/langs/en-US/certd.ts
@@ -471,7 +471,7 @@ export default {
   statusError: "Error",
   actionImportBatch: "Batch Import",
   actionSyncIp: "Sync IP",
-  TitleSyncIp: "Sync IP",
+  modalTitleSyncIp: "Sync IP",
   modalContentSyncIp: "Are you sure to sync IP?",
   notificationSyncComplete: "Sync Complete",
   actionCheckAll: "Check All",
@@ -711,5 +711,19 @@ export default {
   modal: {
     close: "Close",
     viewCertificateTitle: "View Certificate",
+  },
+  domain: {
+    domainManager: "Domain Manager",
+    domainDescription: "used to auto apply for certificate", //管理域名的校验方式，用于申请证书时自动选择验证方式
+    domain: "Domain",
+    challengeType: "Challenge Type",
+    dnsProviderType: "DNS Provider Type",
+    dnsProviderAccess: "DNS Provider Access",
+    httpUploaderType: "HTTP Uploader Type",
+    httpUploaderAccess: "HTTP Uploader Access",
+    httpUploadRootDir: "HTTP Upload Root Dir",
+    disabled: "Disabled",
+    challengeSetting: "Challenge Setting",
+    gotoCnameTip: "Please go to CNAME Record Page",
   },
 };