+++ packages/ui/certd-client/src/locales/langs/zh-CN/certd.ts
@@ -715,4 +715,18 @@ export default {
     close: "关闭",
     viewCertificateTitle: "查看证书",
   },
+  domain: {
+    domainManager: "域名管理",
+    domainDescription: "管理域名的校验方式，用于申请证书时自动选择验证方式",
+    domain: "域名",
+    challengeType: "校验类型",
+    dnsProviderType: "DNS提供商类型",
+    dnsProviderAccess: "DNS提供商授权",
+    httpUploaderType: "上传方式",
+    httpUploaderAccess: "上传授权信息",
+    httpUploadRootDir: "网站根路径",
+    disabled: "禁用/启用",
+    challengeSetting: "校验配置",
+    gotoCnameTip: "CNAME域名配置请前往CNAME记录页面添加",
+  },
 };